Move object-related data structures into separate module
