+++ crates/spartan-farmer/src/object_mappings.rs
@@ -5,7 +5,8 @@ use parity_scale_codec::{Decode, Encode};
 use rocksdb::DB;
 use std::path::Path;
 use std::sync::Arc;
-use subspace_core_primitives::{GlobalObject, Sha256Hash};
+use subspace_core_primitives::objects::GlobalObject;
+use subspace_core_primitives::Sha256Hash;
 use thiserror::Error;
 
 #[derive(Debug, Error)]
